Tox: - code cleaning - removed unneeded dependencies - posible fix for disconnect
git-svn-id: http://svn.miranda-ng.org/main/trunk@12183 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Tox/src/tox_avatars.cpp
@@ -35,7 +35,7 @@ void CToxProto::SetToxAvatar(std::tstring path, bool checkHash)
 		return;
 	}*/
 
-	long length;
+	size_t length;
 	uint8_t *data;
 	FILE *hFile = _tfopen(path.c_str(), L"rb");
 	if (!hFile)
@@ -55,7 +55,7 @@ void CToxProto::SetToxAvatar(std::tstring path, bool checkHash)
 	}
 
 	data = (uint8_t*)mir_alloc(length);
-	long read = fread(data, sizeof(uint8_t), length, hFile);
+	size_t read = fread(data, sizeof(uint8_t), length, hFile);
 	if (read != length)
 	{
 		fclose(hFile);
